add relationship between tasks and users

+++ app.py
@@ -23,7 +23,6 @@ login_manager.login_view = 'login'
 def load_user(user_id):
     return User.query.get(int(user_id))
 
-
 class User (db.Model, UserMixin):
     id = db.Column(db.Integer, primary_key=True)
     username = db.Column(db.String(30), nullable=False, unique=True)
@@ -32,6 +31,7 @@ class User (db.Model, UserMixin):
 class Todo(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     content = db.Column(db.String(200), nullable=False)
+    user_id = db.Column(db.Integer, db.ForeignKey('user.id'), nullable=False)
     date_created = db.Column(db.DateTime, default=datetime.utcnow) 
 
     def __repr__(self):
@@ -90,7 +90,7 @@ def register():
 def index():
     if request.method == 'POST':
         task_content = request.form['content']
-        new_task = Todo(content=task_content)
+        new_task = Todo(content=task_content, user_id=current_user.get_id())
 
         try:
             db.session.add(new_task)
@@ -100,13 +100,16 @@ def index():
             return 'There was an issue adding your task'
 
     else:
-        tasks = Todo.query.order_by(Todo.date_created).all()
+        tasks = Todo.query.filter_by(user_id=current_user.get_id()).all()
         return render_template('index.html', tasks=tasks)
 
 @app.route('/delete/<int:id>')
 @login_required
 def delete(id):
     task_to_delete = Todo.query.get_or_404(id)
+
+    if str(task_to_delete.user_id) != str(current_user.get_id()):
+        return 'You are unauthorised'
 
     try:
         db.session.delete(task_to_delete)
@@ -119,6 +122,9 @@ def delete(id):
 @login_required
 def update(id):
     task = Todo.query.get_or_404(id)
+
+    if str(task.user_id) != str(current_user.get_id()):
+            return 'You are unauthorised'
 
     if request.method == 'POST':
         task.content = request.form['content']
